backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-anthropic
langchain-google-genai
supabase
python-dotenv
apscheduler
resend
requests